// src/app/model/student-model.ts
export interface StudentModel {
    idStudent: string;
    name: string;
    lastName: string;
    documentType: string;
    documentNumber: string;
    gender: string;
    birthDate: string;
    baptism: string;
    communion: string;
    email: string;
    birthPlace: string;
    level: string;
    grade: string;
    section: string;
    status: string;
}